move tests to flutter package
